refactoring and tuning
