+++ exposure-notification/src/main/java/fi/thl/covid19/exposurenotification/MaintenanceService.java
@@ -4,7 +4,6 @@ import fi.thl.covid19.exposurenotification.batch.BatchFileService;
 import fi.thl.covid19.exposurenotification.batch.BatchFileStorage;
 import fi.thl.covid19.exposurenotification.batch.BatchIntervals;
 import fi.thl.covid19.exposurenotification.diagnosiskey.DiagnosisKeyDao;
-import fi.thl.covid19.exposurenotification.efgs.FederationGatewayService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
@@ -13,7 +12,6 @@ import org.springframework.stereotype.Service;
 
 import java.time.Duration;
 import java.time.Instant;
-import java.util.Optional;
 
 import static java.util.Objects.requireNonNull;
 import static net.logstash.logback.argument.StructuredArguments.keyValue;
@@ -26,19 +24,16 @@ public class MaintenanceService {
     private final DiagnosisKeyDao dao;
     private final BatchFileStorage batchFileStorage;
     private final BatchFileService batchFileService;
-    private final FederationGatewayService fgs;
 
     private final Duration tokenVerificationLifetime;
 
     public MaintenanceService(DiagnosisKeyDao dao,
                               BatchFileService batchFileService,
                               BatchFileStorage batchFileStorage,
-                              FederationGatewayService fgs,
                               @Value("${covid19.maintenance.token-verification-lifetime}") Duration tokenVerificationLifetime) {
         this.dao = requireNonNull(dao);
         this.batchFileStorage = requireNonNull(batchFileStorage);
         this.batchFileService = requireNonNull(batchFileService);
-        this.fgs = fgs;
         this.tokenVerificationLifetime = requireNonNull(tokenVerificationLifetime);
         LOG.info("Initialized: {}", keyValue("tokenVerificationLifetime", tokenVerificationLifetime));
     }
@@ -63,25 +58,5 @@ public class MaintenanceService {
                 keyValue("removedVerifications", removedVerifications),
                 keyValue("removedBatches", removedKeys),
                 keyValue("addedBatches", addedBatches));
-    }
-
-    // TODO: rename service or make own service for this
-    @Scheduled(initialDelayString = "${covid19.federation-gateway.upload-interval}",
-            fixedRateString = "${covid19.federation-gateway.upload-interval}")
-    public void syncEfgs() {
-        runExportToEfgs();
-        runImportFromEfgs();
-    }
-
-    private void runExportToEfgs() {
-        LOG.info("Starting scheduled export to efgs.");
-        fgs.startOutbound();
-        LOG.info("Scheduled export to efgs finished.");
-    }
-
-    private void runImportFromEfgs() {
-        LOG.info("Starting scheduled import from efgs.");
-        fgs.startInbound(Optional.empty());
-        LOG.info("Scheduled export to efgs finished.");
     }
 }
